refactor: Path 로직 분리, userDefaultsKey 정의
