+++ Projects/AppFoundation/Sources/Define/UserDefaultsKeyDefine.swift
@@ -1,0 +1,12 @@
+//
+//  StringDefine.swift
+//  AppFoundation
+//
+//  Created by 송하민 on 9/28/24.
+//
+
+import Foundation
+
+public enum UserDefaultsKeyDefine: String {
+  case isAppFirst = "isAppFirst"
+}
